fix integration script config
remove `integration_test` config from truffle, use `development` instead, otherwise dx-contracts migrations break.

reorganize integration scripts
