+++ package.json
@@ -11,7 +11,7 @@
     "deploy-mainnet": "truffle migrate --network mainnet",
     "test:truffle": "run-with-testrpc -l 20000000 'truffle test --network unit_testing'",
     "test": "truffle test --network=unit_testing",
-    "integration-test": "truffle exec ./test/exec.js",
+    "integration-test": "truffle exec ./test/integration/helpers/exec.js",
     "integration-tests": "scripts/integrationTests.sh",
     "migrate": "truffle migrate",
     "migrate-reset": "truffle migrate --reset",
