+++ truffle.js
@@ -28,13 +28,6 @@ module.exports = {
       gas: 6500000,
       gasPrice: 5000000000
     },
-    integration_testing: {
-      host: 'localhost',
-      network_id: '*',
-      port: 8545,
-      gas: 6500000,
-      gasPrice: 5000000000
-    },
     // mainnet: {
     //   provider: new HDWalletProvider(mnemonic, 'https://mainnet.infura.io'),
     //   network_id: '1',
